HOTT-1414: Move to shared polymorphic goods nomenclature serializers
