--- app/serializers/api/v2/quotas/definition/chapter_serializer.rb
@@ -1,11 +1,0 @@
-module Api
-  module V2
-    module Quotas
-      module Definition
-        class ChapterSerializer < GoodsNomenclatureSerializer
-          set_type :chapter
-        end
-      end
-    end
-  end
-end
