--- app/serializers/api/v2/quotas/definition/commodity_serializer.rb
@@ -1,11 +1,0 @@
-module Api
-  module V2
-    module Quotas
-      module Definition
-        class CommoditySerializer < GoodsNomenclatureSerializer
-          set_type :commodity
-        end
-      end
-    end
-  end
-end
