--- app/serializers/api/v2/quotas/definition/heading_serializer.rb
@@ -1,11 +1,0 @@
-module Api
-  module V2
-    module Quotas
-      module Definition
-        class HeadingSerializer < GoodsNomenclatureSerializer
-          set_type :heading
-        end
-      end
-    end
-  end
-end
