--- app/serializers/api/v2/quotas/definition/subheading_serializer.rb
@@ -1,11 +1,0 @@
-module Api
-  module V2
-    module Quotas
-      module Definition
-        class SubheadingSerializer < GoodsNomenclatureSerializer
-          set_type :subheading
-        end
-      end
-    end
-  end
-end
