+++ app/serializers/api/v2/shared/chapter_serializer.rb
@@ -1,0 +1,9 @@
+module Api
+  module V2
+    module Shared
+      class ChapterSerializer < GoodsNomenclatureSerializer
+        set_type :chapter
+      end
+    end
+  end
+end
