+++ app/serializers/api/v2/shared/commodity_serializer.rb
@@ -1,0 +1,9 @@
+module Api
+  module V2
+    module Shared
+      class CommoditySerializer < GoodsNomenclatureSerializer
+        set_type :commodity
+      end
+    end
+  end
+end
