+++ app/serializers/api/v2/shared/heading_serializer.rb
@@ -1,0 +1,9 @@
+module Api
+  module V2
+    module Shared
+      class HeadingSerializer < GoodsNomenclatureSerializer
+        set_type :heading
+      end
+    end
+  end
+end
