+++ app/serializers/api/v2/shared/subheading_serializer.rb
@@ -1,0 +1,9 @@
+module Api
+  module V2
+    module Shared
+      class SubheadingSerializer < GoodsNomenclatureSerializer
+        set_type :subheading
+      end
+    end
+  end
+end
