+++ spec/serializers/api/v2/quotas/quota_definition_serializer_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe Api::V2::Quotas::Definition::QuotaDefinitionSerializer do
+RSpec.describe Api::V2::Quotas::QuotaDefinitionSerializer do
   subject(:serializer) { described_class.new(serializable, options) }
 
   let(:serializable) { create(:quota_definition, :with_quota_balance_events) }
