+++ spec/serializers/api/v2/shared/chapter_serializer_spec.rb
@@ -1,3 +1,3 @@
-RSpec.describe Api::V2::Quotas::Definition::ChapterSerializer do
+RSpec.describe Api::V2::Shared::ChapterSerializer do
   it_behaves_like 'a serialized goods nomenclature', 'chapter'
 end
