+++ spec/serializers/api/v2/shared/commodity_serializer_spec.rb
@@ -1,3 +1,3 @@
-RSpec.describe Api::V2::Quotas::Definition::CommoditySerializer do
+RSpec.describe Api::V2::Shared::CommoditySerializer do
   it_behaves_like 'a serialized goods nomenclature', 'commodity'
 end
