+++ spec/serializers/api/v2/shared/goods_nomenclature_serializer_spec.rb
@@ -1,3 +1,3 @@
-RSpec.describe Api::V2::Quotas::Definition::GoodsNomenclatureSerializer do
+RSpec.describe Api::V2::Shared::GoodsNomenclatureSerializer do
   it_behaves_like 'a serialized goods nomenclature', 'goods_nomenclature'
 end
