+++ spec/serializers/api/v2/shared/heading_serializer_spec.rb
@@ -1,3 +1,3 @@
-RSpec.describe Api::V2::Quotas::Definition::HeadingSerializer do
+RSpec.describe Api::V2::Shared::HeadingSerializer do
   it_behaves_like 'a serialized goods nomenclature', 'heading'
 end
