+++ spec/serializers/api/v2/shared/subheading_serializer_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe Api::V2::Quotas::Definition::SubheadingSerializer do
+RSpec.describe Api::V2::Shared::SubheadingSerializer do
   it_behaves_like 'a serialized goods nomenclature', 'subheading' do
     let(:serializable) { create(:goods_nomenclature) }
   end
